backend/requirements.txt
Django==3.2.3
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
python-dotenv==0.21.0
PyYAML==6.0
Pillow==9.0.0
drf-extra-fields==3.7.0
psycopg2-binary==2.9.3
webcolors==1.11.1